requirements.txt
langchain==0.3.25
langchain-google-genai==2.1.4
langgraph==0.4.1
pdfplumber==0.11.6
pydantic==2.11.4
python-dotenv==1.1.0
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
jinja2==3.1.6
streamlit==1.44.1
